move function to new file to avoid bundling issues
